web/requirements.txt
face_recognition_models>=0.3.0
dlib>=19.7
numpy
Pillow
Flask
redis
